Fixed from size_t to int because wrapping was happening :O. Then added forward and backward substitution. Also added solving systems from cholesky

+++ include/staticarrays/matrix.hpp
@@ -3,40 +3,40 @@
 namespace Static
 {
 
-template <typename T, size_t M, size_t N>
+template <typename T, int M, int N>
 class Matrix
 {
     private:
         T data_[M][N];
 
     public:
-        T& operator()(size_t row, size_t col)
+        T& operator()(int row, int col)
         {
             return data_[row][col];
         }
 
-        T operator()(size_t row, size_t col) const
+        T operator()(int row, int col) const
         {
             return data_[row][col];
         }
 };
 
-template <size_t M, size_t N>
+template <int M, int N>
 using Matrixi = Matrix<int,M,N>;
 
-template <size_t M, size_t N>
+template <int M, int N>
 using Matrixf = Matrix<float,M,N>;
 
-template <size_t M, size_t N>
+template <int M, int N>
 using Matrixd = Matrix<double,M,N>;
 
-template <typename T, size_t M, size_t N>
+template <typename T, int M, int N>
 Matrix<T,M,N> operator+(const Matrix<T,M,N>& A, const Matrix<T,M,N>& B)
 {
     Matrix<T,M,N> result;
-    for(size_t row = 0; row < M; ++row)
+    for(int row = 0; row < M; ++row)
     {
-        for(size_t col = 0; col < N; ++col)
+        for(int col = 0; col < N; ++col)
         {
             result(row,col) = A(row,col) + B(row,col);
         }
@@ -44,13 +44,13 @@ Matrix<T,M,N> operator+(const Matrix<T,M,N>& A, const Matrix<T,M,N>& B)
     return result;
 }
 
-template <typename T, size_t M, size_t N>
+template <typename T, int M, int N>
 Matrix<T,M,N> operator-(const Matrix<T,M,N>& A, const Matrix<T,M,N>& B)
 {
     Matrix<T,M,N> result;
-    for(size_t row = 0; row < M; ++row)
+    for(int row = 0; row < M; ++row)
     {
-        for(size_t col = 0; col < N; ++col)
+        for(int col = 0; col < N; ++col)
         {
             result(row,col) = A(row,col) - B(row,col);
         }
